renames and cleanup
