+++ src/Admin/Billing/Views/OrganizationPlanMigrationCohorts/Index.cshtml
@@ -1,7 +1,7 @@
 @using Bit.Core.Billing.Enums
 @using Bit.Core.Billing.Organizations.PlanMigration.ValueObjects
 @using Bit.Core.Enums
-@model Bit.Admin.Billing.Models.Cohorts.CohortsPagedModel
+@model Bit.Admin.Billing.Models.OrganizationPlanMigrationCohorts.CohortsPagedModel
 @{
     ViewData["Title"] = "Cohorts";
 }
