+++ test/Admin.Test/Billing/Controllers/OrganizationPlanMigrationCohortsControllerTests.cs
@@ -1,5 +1,5 @@
 using Bit.Admin.Billing.Controllers;
-using Bit.Admin.Billing.Models.Cohorts;
+using Bit.Admin.Billing.Models.OrganizationPlanMigrationCohorts;
 using Bit.Core;
 using Bit.Core.Billing.Organizations.PlanMigration.Entities;
 using Bit.Core.Billing.Organizations.PlanMigration.Enums;
